Rename to unitMiniSearch for clear intent.
